yrbss.py -> noextract_collections.py
